administrativeregioncodespider/requirements.txt
requests==2.22.0
lxml==4.3.3
Flask==1.1.1
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
fake-useragent==0.1.11


